Remove positional format parameters, use standard C++ specifiers
Replace all POSIX positional parameters (%1\$zu, %1\$d, %1\$s) with standard
C/C++ format specifiers across all 6 translation files.

Translation string changes:
- %1\$zu → %llu (size_t values)
- %2\$zu → %llu (second size_t)
- %1\$d → %d (int values)
- %2\$d → %d (second int)
- %1\$ld → %ld (long values)
- %1\$s → %s (strings)
- %2\$s → %s (second string)

Code changes:
- Simplified Translator::Format() - removed ConvertFormatString()
- Added static_cast<unsigned long long>() for .size() calls
- Use .c_str() for string parameters (stable pointer)

Files:
- src/common/Translator.hpp
- src/taxonomy-tree/TreeStrings_{EN,LT,ES}.cpp
- src/taxonomy-dict/DictStrings_{EN,LT,ES}.cpp
- src/taxonomy-tree/TreeFrame.cpp
- src/taxonomy-tree/TaxonomyData.cpp

+++ src/taxonomy-tree/TaxonomyData.cpp
@@ -71,7 +71,7 @@ bool TaxonomyData::LoadArchive(const wxString& archivePath, std::function<void(c
     }
 
     auto fileList = m_zipReader->GetFileList();
-    updateProgress(10, TR_TREE_FMT(TreeStringId::ProgressArchiveContainsFiles, fileList.size()));
+    updateProgress(10, TR_TREE_FMT(TreeStringId::ProgressArchiveContainsFiles, static_cast<unsigned long long>(fileList.size())));
 
     updateProgress(30, TR_TREE(TreeStringId::ProgressLoadingTaxonomyData));
     if (!LoadTaxonomyData(progressCallback))
@@ -85,7 +85,7 @@ bool TaxonomyData::LoadArchive(const wxString& archivePath, std::function<void(c
 
     updateProgress(80, TR_TREE(TreeStringId::ProgressScanningLanguageFiles));
     std::vector<wxString> languageFiles = m_zipReader->GetLanguageFiles();
-    updateProgress(85, TR_TREE_FMT(TreeStringId::ProgressFoundLanguageFiles, languageFiles.size()));
+    updateProgress(85, TR_TREE_FMT(TreeStringId::ProgressFoundLanguageFiles, static_cast<unsigned long long>(languageFiles.size())));
 
     m_availableLanguages.clear();
     for (const auto& file : languageFiles)
@@ -165,7 +165,7 @@ bool TaxonomyData::LoadTaxonomyData(std::function<void(const wxString&, int)> pr
         return false;
     }
 
-    updateProgress(65, TR_TREE_FMT(TreeStringId::ProgressParsedEntries, tempMap.size()));
+    updateProgress(65, TR_TREE_FMT(TreeStringId::ProgressParsedEntries, static_cast<unsigned long long>(tempMap.size())));
 
     // Convert to TaxonNode map
     m_taxonomyMap.clear();
